Move DumpAST rule to separate 'test' ruleset module.
This will allow to remove inderect JUnit dependency from CLI.
