+++ ktlint-ruleset-test/gradle.properties
@@ -1,0 +1,4 @@
+GROUP=com.pinterest.ktlint
+POM_NAME=ktlint-ruleset-test
+POM_ARTIFACT_ID=ktlint-ruleset-test
+POM_PACKAGING=jar
