+++ ktlint-ruleset-test/src/main/kotlin/com/pinterest/ruleset/test/TestRuleSetProvider.kt
@@ -1,0 +1,11 @@
+package com.pinterest.ruleset.test
+
+import com.pinterest.ktlint.core.RuleSet
+import com.pinterest.ktlint.core.RuleSetProvider
+
+public class TestRuleSetProvider : RuleSetProvider {
+    override fun get(): RuleSet = RuleSet(
+        "test",
+        DumpASTRule(),
+    )
+}
