+++ ktlint-ruleset-test/src/main/kotlin/com/pinterest/ruleset/test/internal/Color.kt
@@ -1,11 +1,11 @@
-package com.pinterest.ktlint.test.internal
+package com.pinterest.ruleset.test.internal
 
 /**
  * Stripped down version of https://github.com/ziggy42/kolor (ziggy42/kolor#6).
  */
-fun String.color(foreground: Color) = "\u001B[${foreground.code}m$this\u001B[0m"
+internal fun String.color(foreground: Color) = "\u001B[${foreground.code}m$this\u001B[0m"
 
-enum class Color(val code: Int) {
+internal enum class Color(val code: Int) {
     BLACK(30),
     RED(31), GREEN(32), YELLOW(33), BLUE(34), MAGENTA(35), CYAN(36),
     LIGHT_GRAY(37), DARK_GRAY(90),
