+++ ktlint-ruleset-test/src/main/resources/META-INF.services/com.pinterest.ktlint.core.RuleSetProvider
@@ -1,0 +1,1 @@
+com.pinterest.ktlint.ruleset.test.TestRuleSetProvider
