+++ settings.gradle
@@ -34,4 +34,5 @@ include ':ktlint-reporter-plain'
 include ':ktlint-ruleset-experimental'
 include ':ktlint-ruleset-standard'
 include ':ktlint-ruleset-template'
+include ':ktlint-ruleset-test'
 include ':ktlint-test'
